+++ WordLite/querywidget.h
@@ -2,12 +2,51 @@
 #define QUERYWIDGET_H
 
 #include <QWidget>
-#include <wordwidget.h>
+#include <QLayout>
+#include <QGroupBox>
+#include <QLabel>
+#include <QPushButton>
+#include <QScrollArea>
+#include <QMenu>
+#include <QActionGroup>
+#include <QAction>
+#include <QInputDialog>
+#include <QLineEdit>
+#include <QMessageBox>
+#include <QToolButton>
+#include <QWidgetAction>
+#include <QCheckBox>
 #include <./back_head/database.h>
+#include <./API/dictionaryapi.h>
 
 namespace Ui {
 class QueryWidget;
+class WordWidget;
 }
+
+class WordWidget : public QWidget
+{
+    Q_OBJECT
+
+public:
+    explicit WordWidget(QString word, WordDatabase* ptr, bool& useapi, QWidget *parent = nullptr);
+    ~WordWidget();
+
+    QString word;
+    bool useAPI = false;
+    WordDatabase* DBptr;
+
+    // 设置UI和连接信号槽
+    void setupUI(QVector<Word> wordList);
+    void connectSignals();
+
+private: signals:
+    void sendMes(QString message, int timeout=0);
+    void clearMes();
+private:
+};
+
+
 
 class QueryWidget : public QWidget
 {
@@ -19,6 +58,8 @@ public:
 
     WordWidget* wordWidget;
     WordDatabase* wordDataBase;
+    DictionaryAPI* dictionaryAPI;
+    bool useAPI = false;
     QString defaultWordDB;
     QVector<QString> wordDBList;
 
@@ -32,8 +73,9 @@ private slots:
 
     void on_comboBox_currentTextChanged(const QString &arg1);
 
-    void showMes(QString message, int timeout);
+    void sendMess(QString message, int timeout);
     void clearMess();
+
 
 private: signals:
     void sendMes(QString message, int timeout=0);
@@ -43,6 +85,9 @@ private: signals:
 private:
     Ui::QueryWidget *ui;
 };
+
+
+
